requirements.txt
Pillow==9.4.0
opencv-python==4.7.0.72
torch==1.13.1
torchvision==0.14.1
kornia==0.6.10
pyYAML==6.0
scipy==1.10.1
ray==2.3.1
tabulate==0.9.0 # TrackEval
